(refactor) add get_in_range + split up candle model into files
