+++ lib/models/candle/get_in_range.js
@@ -1,0 +1,5 @@
+module.exports = ({ getAll }) => ({ type, symbol, tf } = {}, { start, end }) => (
+  getAll({
+    type, symbol, tf
+  }).filter(({ mts }) => mts >= start && mts <= end)
+)
