+++ lib/models/candle/sync_range.js
@@ -1,0 +1,3 @@
+module.exports = (methods) => ({ type, symbol, tf, start, end } = {}) => {
+
+}
